requirements.txt
Django>=5.2.0,<5.3.0
Pillow>=10.0.0
gunicorn>=21.0.0
psycopg2-binary>=2.9.0
python-decouple>=3.8
whitenoise>=6.5.0
dj-database-url>=2.1.0
django-ninja>=0.24.0
pydantic>=1.10.0
django-ninja-jwt>=0.1.0